update for realtime
Former-commit-id: 3e1ff80db137cffb0343e2cad3474d5e08271d7f
Former-commit-id: 0a69f47888ef888e7e6282f6f5c50a00b5bfb2c1 [formerly e91d7ad1b1d132eee67934f6af9de640d00d0031]
Former-commit-id: 5a47683c25793237f61426008558ba989a1f52fa
Former-commit-id: d201d4c973add25b0bb880709b79be97e63a968a

+++ charts2/android/src/akylas/charts2/datasets/BarChartDataSetProxy.java
@@ -56,7 +56,7 @@ public class BarChartDataSetProxy
     }
     
     @Override
-    protected BarEntry dataEntryFromNumber(Number number, int index) {
+    protected BarEntry dataEntryFromNumber(Number number, float index) {
         BarEntry entry = (BarEntry) super.dataEntryFromNumber(number, index);
         if (entry != null) {
 //            entry.setHigh(number.floatValue());
